WIP: trying a major performance optimization for bertscore
Before, I was running score(...) for each row separately. it seems like there's a lot of overhead in loading the model in memory. passing a list of all candidates + references per model has a massive speedup -- on 10 examples, around 45 seconds with the old approach to 15 seconds with the new approach. that should shorten run time from around 8 hours down to 2 hours, so we can even try a couple different models and see what is giving best subjective performance. still need to implement as a function and in a script